2020.05.22

+++ Cpp_Algorithmic_Problem_Solving/src/001.cpp
@@ -13,26 +13,26 @@ void p001() {
 
     for(int tries = 0; tries < T; tries++) {
 
-        cin >> D >> S;
+        cin >> D >> S; // number of all days, and number of days that we count continously
 
         int* Dates = new int[D];
 
-        for(int i = 0; i < D; i++) {
+        for(int i = 0; i < D; i++) { // getting costs of dates
             cin >> Dates[i];
         }
         
-        double min = 100;
+        double min = 100; // minimum cost. it initialized with maximum possible value.
 
-        for(; S <= D; S++) {
-            for(int i = 0; i + (S - 1) < D; i++) {
+        for(; S <= D; S++) { // looping by increase the number of continous days.
+            for(int i = 0; i + (S - 1) < D; i++) { // set each starting point of continous days.
                 double result = 0;
                 for(int j = 0; j < S; j++) {
-                    result += Dates[i + j];
+                    result += (double)Dates[i + j]; 
                 }
-                result = result / S;
+                result = result / S; // getting average cost.
 
                 if(result < min) {
-                    min = result;
+                    min = result;   // if average cost per day is smaller than minimum cost before, switch it.
                 }
             }
         }
